fetch関数整理
